+++ resolwe_bio/tests/workflows/test_rna_seq.py
@@ -1,6 +1,12 @@
 # pylint: disable=missing-docstring
-from resolwe_bio.utils.test import BioProcessTestCase
+from django.contrib.auth.models import AnonymousUser
+from django.test import LiveServerTestCase
+
+from guardian.shortcuts import assign_perm
+
+from resolwe_bio.utils.test import BioProcessTestCase, TransactionBioProcessTestCase
 from resolwe.flow.models import Data
+from resolwe.test.utils import with_resolwe_host
 
 
 class RNASeqWorkflowTestCase(BioProcessTestCase):
@@ -128,6 +134,56 @@ class RNASeqWorkflowTestCase(BioProcessTestCase):
             'trailing': 1,
             'stranded': 'no',
             'id_attribute': 'transcript_id'
+        })
+
+        for data in Data.objects.all():
+            self.assertStatus(data, Data.STATUS_DONE)
+
+        workflow = Data.objects.last()
+        self.assertFile(workflow, 'rc', 'workflow_rnaseq_paired_rc.tab.gz', compression='gzip')
+        self.assertFields(workflow, 'exp_type', 'TPM')
+        self.assertFields(workflow, 'source', 'DICTYBASE')
+
+
+class RNASeqDSSTestCase(TransactionBioProcessTestCase, LiveServerTestCase):
+    @with_resolwe_host
+    def test_rnaseq_dss(self):
+        single_reads = self.prepare_reads()
+        paired_reads = self.prepare_paired_reads()
+        genome = self.prepare_genome()
+        genome.slug = 'genome-mm10'
+        genome.save()
+        assign_perm('view_data', AnonymousUser(), genome)
+        annotation = self.prepare_annotation('annotation.gtf.gz')
+        annotation.slug = 'annotation-mm10'
+        annotation.save()
+        assign_perm('view_data', AnonymousUser(), annotation)
+        adapters = self.prepare_adapters()
+        adapters.slug = 'adapters-illumina'
+        adapters.save()
+        assign_perm('view_data', AnonymousUser(), adapters)
+
+        self.run_process('dss-rna-seq', {
+            'genome_and_annotation': 'mm',
+            'reads': single_reads.id,
+            'id_attribute': 'transcript_id',
+            'adapters': 'no'
+        })
+
+        for data in Data.objects.all():
+            self.assertStatus(data, Data.STATUS_DONE)
+
+        workflow = Data.objects.last()
+        self.assertFile(workflow, 'rc', 'workflow_rnaseq_single_rc.tab.gz', compression='gzip')
+        self.assertFields(workflow, 'exp_type', 'TPM')
+        self.assertFields(workflow, 'source', 'DICTYBASE')
+
+        self.run_process('dss-rna-seq', {
+            'genome_and_annotation': 'mm',
+            'reads': paired_reads.id,
+            'trailing': 1,
+            'id_attribute': 'transcript_id',
+            'adapters': 'no'
         })
 
         for data in Data.objects.all():