+++ resolwe_bio/tools/rnaseq.py
@@ -28,16 +28,17 @@ def parse_arguments():
     parser.add_argument('minlen', type=int, help='Min length')
     parser.add_argument('trailing', type=int, help='Trailing quality')
     parser.add_argument('stranded', help='Stranded')
+    parser.add_argument('id_attribute', help='ID attribute')
     return parser.parse_args()
 
 
-def get_data_id(res, slug, type=None):
-    """Throw an error if data object is missing."""
-    data_objs = res.data.filter(slug=slug, type=type)
-    if not data_objs:
-        print("Server configuration error: no slug '{}' of type '{}'.".format(slug, type))
+def get_data_id(res, slug, type_=None):
+    """Get id of a data object from slug."""
+    try:
+        return res.data.get(slug=slug, type=type_).id
+    except:
+        print("Server configuration error: no slug '{}' of type '{}'.".format(slug, type_))
         exit(1)
-    return data_objs[0].id
 
 
 def main():
@@ -48,7 +49,8 @@ def main():
             'reads': args.reads,
             'minlen': args.minlen,
             'trailing': args.trailing,
-            'stranded': args.stranded
+            'stranded': args.stranded,
+            'id_attribute': args.id_attribute
         }
     }
     if args.type.startswith('data:reads:fastq:single:'):
@@ -61,7 +63,8 @@ def main():
     res = Resolwe()
     data['input']['genome'] = get_data_id(res, gna['genome'], 'data:genome:fasta:')
     data['input']['annotation'] = get_data_id(res, gna['annotation'], 'data:annotation:gtf:')
-    data['input']['adapters'] = get_data_id(res, ada, 'data:seq:nucleotide:')
+    if ada:
+        data['input']['adapters'] = get_data_id(res, ada, 'data:seq:nucleotide:')
     print('run {}'.format(dumps(data, separators=(',', ':'))))